Change in project License from AGPL 3.0 to Apache 2.0 prior to first public release.
Several factors influenced this decision, the largest being community building and making things as easy as possible for folks to get started with the project.

We don't however believe Open Source == Free and will continue to investigate the best way to commercialize this IP, restrictive copy-left licenses aren't however the answer.