added grails javascript providers from their plug-ins to nimble.
Do not know if this is what should be done, but I am not keen on the idea of adding more plug-in dependencies to nimble.
This also makes the js libraries used in nimble stay fixed to what ever nimble has for that particular version.
These can be stripped later, if this is deemed unnecessary/unwanted.